[tee][api] Clarify undefined behavior of memory functions
The spec does not define the behaviour of TEE_MemMove, TEE_MemCompare,
or TEE_MemFill when the provided pointers are null (ditto for the
related libc functions). However, it is a precondition of the Rust
standard library functions that we implement these with that these
pointers are non-null. For clearer panicking and better interface
hygiene, we add non-nullilty asserts before these calls.

In the case of TEE_MemMove, we actually go further by treating a size of
zero as a no-op, even when either of the provided pointers is null! It
turns out that there are xtest cases that explicitly rely on being able
to do this! Given that, and the fact that this isn't against the letter
of the spec, we deem this behavior as sensible for our implementation.

Bug: 371213067
Change-Id: I24be812563ac9eef0eba24554d16265e13d6ddf8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1138792
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>