[Honeydew] Move system_power_state_controller affordance related code into one place.
Doing this as part of enabling self-servicable Honeydew user affordance
contributions.

Current Honeydew code structure organizes the code based on interfaces,
implementation, unit-tests, functional-tests etc. As a result code
related to one affordance is in multiple sub folders inside Honeydew.
This makes it difficult to give a complete OWNERShip of a user
affordance to corresponding user team implementing that affordance. So
re-organize the folder structure such that all the code associated with a
given affordance (or transport or auxiliary device etc) is at one place
with a facility to add OWNER, METADATA.textproto files if needed.

At a high level, current folder sturcutre looks similar to
https://paste.googleplex.com/6465845936259072 and new folder strucurtre
will be similar to https://paste.googleplex.com/6084362981605376

This will be done in multiple phases. As part of this CL, we are
updating system_power_state_controller affordance.

Change-Id: I0a1f1c4db14b1834f2f51fef394e8c9f3f567a87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135917
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Kevin Cho <chok@google.com>
Commit-Queue: Pavan Kumar Juturu <jpavankumar@google.com>