[ffx] In strict mode, ignore env variables
Rather than inspecting the entire config to look for env variables
(which requires the script to pass in config variables that may be
unused by the subtool), we should just ignore values that refer
to env vars, relying on the failure to find a required config entry
to produce a normal error.

For instance, with this change, `ffx ... --strict target echo`
will produce an error noting that `ssh.priv` is unset, but will not
produce errors about any other env vars.

Change-Id: Iddf2d35ac58f9bf30e119cd80af3e2cb5a111a7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139613
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>