[sysmem] add clarifying comment re. allocation
Add a comment clarifying why we don't wait for old collections to be
fully gone before attempting allocation of a new collection, and point
toward AttachLifetimeTracking in case a client wants to wait for any
old collections known to that client.

The docs are also separately updated to clarify this as well.

BUG: b/42150211
Change-Id: Ib006d90847f9f3b8c5a3603f85c9f718d5b67e5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136614
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>