[sestarnix] Add Starnix kernel runner option for fs seclabels
Allow an SELinux Security Context to be specified in the manifest
of a Starnix component, for the Starnix container to apply to all
files in all filesystems mounted for that component.

The new "fsseclabel" argument may now be provided in the "program"
section, alongside the existing "seclabel" argument that controls
the label assigned to the component's initial task.

Bug: 373819089
Change-Id: I64f2831028107f7f0d8de26316677f7a186e36ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139852
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>