[netstack3] Remove rules after table removal
This change prepares for installing rules in Core. In Core, a rule
can hold a strong reference to the table, so we need to make sure
when a table is removed, all rules that reference it should also
be removed.

Bug: 357858471
Change-Id: Ie6b04ca06320194ff1163170612ea6b5267a7038
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139580
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>