[assembly] Allow overrides to __append_to_ lists
The default merge strategy for developer overrides is to fully override
any key found in a dictionary. This is simple to understand but causes
extra toil for developers that want to add a new entry to a list. It
would require the developer to find the config definition, copy the
previous list into their developer overrides, then add their entry to
the list. The config definition may be in a separate repository at a
specific pinned version, therefore the previous list could be hard to
find.

This CL adds the ability to developers to indicate that they want to
append to the previous list instead of overriding it.

We cannot use periods or colons as delimiters, because this must work
with both GN and Bazel. Extra underscores are our best option.

Suggest we have this config:

{
  key: [ value1, value2, value3 ],
}

Developers can put this in their overrides to append:

{
  __append_to_key: [ value4 ],
}

And the result would be:

{
  key: [ value1, value2, value3, value4 ]
}

Change-Id: I516a15048a5ac81de6ed86e8973a84b77a1b7a74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137515
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>