[fdf] Create a quarantine node shutdown intent
This intent will put the node into a quarantine state after it fails to
start a driver, and stops+removes the failed-to-start driver.

This way disabling the driver that had failed will cause it to go
to the orphan list as part of the rebind, and then be available for new
drivers to match with.

This also makes our ffx tool print out "(quarantined)" with the driver if
the node is quarantined.

Tests: Created a new integration test, 'replacefaileddriver_test'

Bug: 370048627

Change-Id: I1afa7758938fb68c97476611ebad59b19826b528
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127478
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>