[cpu_profiler] Increase Resilience when Starting
Between configuring and starting, we may lose one or more of our targets
because they exited. This patch decreases previous error when a process
or job was not found to a warning and allows us to carry on if we fail
to attach to a target.

Bug: 371463369
Multiply: cpu_profiler.sh

Change-Id: I5e8d57fd5f8c51c621bf2803c0d15ed370d92b78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1138265
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>