[profiler] Tolerate ZX_ERR_BAD_STATE when working with tasks.
This error code typically means that the task has already exited and
isn't able to be attached to the sampler or receive an exception
channel. It's common for this to happen with short-lived tasks
and is hit much more frequently when doing system-wide profiles.

Bug: 359658592

Change-Id: I0974e4ea3448ea2a33d87cab1aef68df80675d55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139576
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>