[debugger] Add job-only flag to "attach"
This adds the ability to specify job-only filters from the "attach"
command. On Fuchsia, this will cause any matches for this filter to
attach directly to the job of any matches. Used in conjunction with the
"weak" option, this attaches to a job's Debugger exception channel. When
"weak" is not specified, this attaches to the standard exception
channel, putting us on the path to receive exception notifications
without being attached directly to the process that caused the
exception.

Specifying this option on linux will result in an error.

BUG: 362240229
BUG: 362240382
Change-Id: Ibe0f99d8f005a6e50fb3f7a204a53a12103f3e94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139994
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>