[starnix] Always clear kernel signal on adb read
Normally for hanging gets we don't raise the signal until the next
hanging get has been scheduled. For adb reads we don't know when that
will come, and if the connection gets reset between reads the proxy will
get stuck waiting for the kernel signal.

The current code still mostly worked because the hang would not block
the new proxy that was created for the re-established connection.

However, if the connection was ever reset in this state it would cause
all future suspend requests to succeeed but immediately get woken up, as
if an adb command had come in.

Bug: 373736681

Change-Id: Ifa5a4e9af3b84f24fec84203467ca354293f6f03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140792
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>