[drivers][rust] Improve incoming ergonomics
Implements library support for connecting to protocols and services in
the driver's incoming namespace more easily.

In the process, this adds some new functionality to `fuchsia-component`
libraries to connect to a service at an arbitrary root directory.

It also changes the `IterablePath` trait to use an impl return trait
instead of a boxed dyn in `cm_types` since that's possible now and
wasn't when it was originally written.

Fixed: 373918951
Change-Id: Ibcbcad077c26abd6c0e57d3279bb61c972a47679
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136552
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>