[thread_sampler] Handle attaching to Lifecycle:INITIAL threads
Right now, the kernel thread sampler will refuse to attach to threads
that are still initializing, dead, or dying, returning ZX_ERR_BAD_STATE
for all of them. This makes it difficult for a user to distinguish
between trying to attach to a thread that will not give samples, versus
attaching to a thread that will not give samples /yet/.

This allows the sampler to enable sampling on a thread that's not yet
running so that when it does start running, it's eligible to have
samples taken.

Fixed: 360927412
Fixed: 374109518
Change-Id: Id8c8c421700abeadbfe7e4c95b6cc5014b98bdd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140212
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
API-Review: Adam Barth <abarth@google.com>