[flake][driver-host] Fix Start_ReturnError flake
This was flaking because in certain instances the driver's shutdown
handler where we quit the async loop, was running earlier than the
ResetQuit happening on the test side, where the quit there was to signal
completion of the start async call. This changes the start to use a sync
completion rather than the loop's quit so that we don't reset the quit
state being set from the shutdown side.

Multiply: driver-host2-test

Bug: 374186610
Change-Id: Ia48e79d6fa66b81cd16bfd504b7b681445663a4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140643
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>