[fidl][build][sdk] Limit SDK categories in documentation
Previously, any FIDL library that specified an `sdk_category` would be
included in the generated and published documentation if its `..._sdk`
target appeared in the build tree. Rather than relying on appropriate
use of the targets, specify which categories should be published. This
should not change the set of libraries for which documentation is
published as long as the `..._sdk` targets are only depended on by
targets in `//sdk/fidl/BUILD.gn`.

This replaces a condition that was redundant / always true.

Bug: 366221510
Change-Id: Ie3c2cd6d345602cd90177e8ec7a6d1527a403ffb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140108
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>