[debug-dash-launcher] Use the bootfs package resolver
The launcher now uses the recently exposed bootfs package resolver to
resolve fuchsia-boot:// package URLs.

The tools trampolines are stored in directories named after the full
URL (with / replaced with | and : replaced with ^) instead of just the
first segment of the path now that it is easier to have packages with
the same path but different scheme, host, etc. This is user visible
but clients should be using the tools binaries via $PATH which is
unchanged.

Fixed: 329167311
Change-Id: If25c51422fea33ca1dab09cb8ff954d7123ae9ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137867
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>