[sdk] Create //sdk:bazel_internal_only_idk
The `//sdk:platform` collection is renamed to
`//sdk:bazel_internal_only` to clarify that its content
is only used by the in-tree Bazel `@internal_sdk`
repository.

This also creates `//sdk:bazel_internal_only_idk` to
wrap it as a proper IDK, instead of a collection for
consistency.

+ Fix a typo in an assert() in build/sdk/idk.gni
+ Fix some LINT ThenChange paths.

Change-Id: I46288a420ba95151deb268081bf9c1a438eab141
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139999
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>