[sdk] Move //build/bazel/bazel:idk to //sdk:bazel_in_tree_idk
This renames and changes the location of the IDK that is built
to populate the content of the @fuchsia_sdk repository in
the platform build's internal Bazel workspace.

The new name is clearer about its intended usage, and the
next location ensures that all IDKs are defined in //sdk/BUILD.gn
which makes things more consistent.

There are still some extra validation targets under
//build/bazel/bazel_sdk, but I prefer to let them here since
they involve both GN and Bazel graphs.

There is a new //sdk:bazel_in_tree_idk_tests target that
runs the Bazel SDK test suite directly against this IDK.

It is not necessary to add this to regular tests, as we
already run the test suite on the @fuchsia_sdk repository
itself, which is what really matters for the build.

+ Rename `@fuchsia_idk` repository to `@fuchsia_in_tree_idk`.

Bug: 367274537
Change-Id: Ib6b6300b407eab9908a991349a6851de80244236
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140000
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>