[profile] Record FX_BUILD_UUID in trace metadata
This will be used to assure that we have merged data
for the same build (which may be scattered across different
trace logs).

The metadata is optional and can be empty, so the profile_wrap.sh
script can be used for non-build purposes.

Factor out useful functions/patterns to trace_tools.py.
JSON traces are semi-formatted: left indentation is correct,
but trace event lines are kept on a single line without wrapping.

Bug: 370569880
Change-Id: I7eb0d91be3968aeb4f411c7c2667b2b38b390973
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141352
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Fang <fangism@google.com>