[kernel][rppm] Introduce PowerLevelController
Introduced as an interface to keep the rest of the code working on
userspace. This is just an abstraction over a port.

Intoduced kernel registry, as a wrapper on top of the power domain
registry which provides a mutex over it. Next CL will wire syscalls up
and introduce proper core tests.

Unit test just checks that the port post packets when the port has
active handles, and when attempting to queue on a port with no active
handles it will return an error and set the is_serving flag to false.

The scheduler can check this flag, when to see whether it should
consider active power states or not. This flag can later be set,
if we observe no updates coming in a per domain basis.

domain_->controller()->is_serving();

Bug: 330349572
Fixed: 374124659
Change-Id: I503f90a94a1325ffdb5823e771ca72d72654b372
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095738
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>