[package-tool] Deflake tests
The product bundle publishing test implicitly assumed that the
modification timestamps of the packages in the product bundle and its
counterpart in the written repo would end up with a timestamp within the
same second. This needs to be relaxed in case the test runs slowly or
crosses just into the next second while running.

Bug: b/372116742
Change-Id: Ia34322dce6212b9d12b069bf640eced029df895e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141874
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>