[ffx][ota] Adding product bundle option to target update
This allows the specifying of a product bundle when running
`ffx target update check-now|force-install`.

By default, these commands communicate with the registered package
server, which may or maynot be serving the desired update package.

This feature allows the user to specify to use a product bundle to
provide the update package. This is either done with an explicit path
to a product bundle directory, or if the path is not specified, the
path is retrieved from the configuration using the key `product.path`.

The implementation of this feature then starts a package server,
registers it with the device, then performs the update.

Bug: 364376140
Change-Id: Ife8895ae588ddd284da6987ad4cd2ac0cfaadda1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130267
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>