[kernel] Introduce power_management::ControllerDpc
This is a workaround for escaping scheduler context when requesting a
power level transition from the Cpu Driver.

The Scheduler provides the storage for both the Dpc task and the Timer
object, such that the objects are guaranteed to outlive the timer and
the DPC. This means that whenever a Dpc is queued, the scheduler wont be
able to request a new power level transition, thought It may choose to
just overwrite any existing pending transition.

When the Timer fires, it will queue a Dpc task, which when executed will
take ownership and reset the pending transition from the scheduler. This
means that until such moment happens, the scheduler is free to
update/change the pending transition or even cancel. The scheduler can
verify whether a transition was posted or not, if they have pending
transitions when it needs to rescheduler again.

The operation mode is when the scheduler decides to switch to a given
level it will:
+ Stash the transition provided by PowerState::RequestTransition
+ Call ControllerDpc::NotifyPendingUpdate

Bug: 330349572, 347701043
Change-Id: Ibf1aa820b9b1cd0593e2562c2608a88e367d7a23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137756
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>