[audio][adr] Wait for initial device detection
Ensure that preexisting devices are detected and initialized
before responding to `Registry/WatchDevicesAdded`.

In practice this is not a requirement today because the
ADR is eagerly started on products where it is included.
However, this will unblock certain testing scenarios, and
allows ADR to instead be started on demand if desired.

Bug: b/359344835

Change-Id: I7f1859480df5e1807811504bada42478fd53d63c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141292
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>