[build][sdk] Internal doc for IDK Bazel migration.
Provide a documentation describing the various parts, related
to IDKs and SDKs, that the platform build needs to manage,
and how they are going to evolve.

This document will evolve as migration progresses but is
useful to prepare interested developers to the upcoming
changes.

Bug: 367274537
Change-Id: I977912d97aebcfcf1e32d4e75b31bcb47f018298
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141372
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>