[netstack3] Retain conntrack entry across loopback
Currently all packets look up their conntrack entry when they hit the
INGRESS hook. This causes problems for packets that are NATed and then
looped back: they may no longer match the original or reply tuples of
their connection, so it cannot be retrieved, which results in a new
conntrack entry erroneously being created.

Instead, hold on to a weak reference to a packet's conntrack entry in
the loopback RX and TX queues' per-packet metadata. This allows
looped-back packets to hold onto their conntrack entry on the way out
and for the IP layer to retrieve it on ingress, and avoid allocating a
new connection. This also happens to allow us to drop packets when
finalizing a connection fails (https://fxbug.dev/372543386), something
that, without this change, is not possible to do without breaking NAT
over loopback.

Fixed: 372545312
Test: netstack3-filter-test
      netstack3-core-ip-integration-test
      netstack-filter-fidl-tests
Change-Id: Ib57e184c22abed1b022adabe4349caded78eb096
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137754
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>