[netcfg] Remove masquerade config on close
Ensure that when a `fuchsia.net.masquerade` controller disconnects, it's
masquerade config is removed. This upholds the documented behavior on
the API, that the "protocol is associated with the lifetime of the IP
masquerade".

While here, downgrade some `error!` calls to `warn!`. If disconnect
races with interface removal (which happens from time to time in our
integration test teardown), the disconnect logic will observe `NotFound`
when disabling the masquerade configuration. With `error!` this lead to
spurious test failures.

Multiply: policy-management-integration-test

Fixed: 374285849
Change-Id: I439bd18f9d69952b67ef9323c1b6e54a955f4a4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140772
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>