[bazel] Add //build/bazel:generate_final_in_tree_idk
This adds a new GN target to build the final export version
of the in-tree IDK (using Bazel), and creating a symlink to
the resulting output directory.

This allows running the Fuchsia Bazel SDK test suite against
it directly, as shown by the instructions added to its README.md file.

This is only useful for local debugging during the Bazel migration.
Nothing needs to depend this.

Bug: 367274537
Change-Id: I9725a2909f267232e3220f20f17dcbd1148ac2a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140256
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>