[zxio] Remove attribute equality checks.
They had a few small bugs in them due to missing fields that have since
been added. It was only being used as part of attr_set to ensure that
extra "has" flags weren't being passed in. This deletes a long chunk of
human-managed code that we keep forgetting to update properly and
replaces it with an iteration over the struct of bools as a pointer
array.

Change-Id: I082062304949a08bd88db80e6ea86c07d595e150
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135513
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Nikita Jindal <nikitajindal@google.com>