[hrtimer][vim3] Fix Stop API IRQ handling
Remove calls to cancel IRQ handlers in the FIDL Stop() API
and starts them at initialization.

Canceling an IRQ handler can race with an incoming IRQ,
which could then become pending and never be acknowledged.
The reason the hrtimer canceled IRQ handlers was to
prevent an incoming IRQ from generating a notification after
the FIDL Stop() API was called.

Hence, we change the FIDL Stop() API removing the call to
Cancel() and we update the IRQ handler method such that
it acts only on started timers and valid IRQs.
If we get an IRQ handler call after we have stopped the timer
we'll log and ack this IRQ but not attempt to restart the timer
(normally the IRQ handler will not be called since Stop() stops
the timer hardware).

For the Cancel() on shutdown, added a warning on error and
a TODO for a generalized solution to lost IRQs in that case.

Also improved logging and removed unused code.

Fixed: 369886005
Test: fx test -o aml-hrtimer-test
Multiply: aml-hrtimer-test

Change-Id: I0c8d5c5753c3dc7157f8b030bb2aad025ede4bdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141433
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>