[go] The go runtime and fidl-enabled go is fuchsia-only
This is step 2 of 3 of a restricting of the Golang Runtime to only
fuchsia binaries, as the last remaining host binaries are tests that
are also run on fuchsia.

This restricts the use of the fuchsia go runtime to just fuchsia
binaries, and removes the last host-side tests that were using
the fuchsia Go runtime.

Change-Id: I6a2c926fe339339dd6be2b8d964b2b26adf46084
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141944
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Turner <digit@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>