[starnix][vfs] Restrict xattrs outside of user namespace.
We previously restricted access for non-file and non-dir types for the
user namespace, and restricted access to the security namespace. We did
not prevent others from using every other namespace which meant that
they could access them all.

The real intention here is that symlinks are used with permissions 777
which would allow anyone accessing a symlink to attach xattrs to it.
This is the reason the "user" namespace is only allowed on regular files
and directories, it presumes that all other namespaces require special
privileges.

This will notably prevent use for non-CAP_SYS_ADMIN users from accessing
the "system" and "trusted" namespaces that may often be used, but I
don't think we're supporting anything that does that right now.

Change-Id: I6ea33dec83b7712144acf28dbba721e1598629dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143212
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>