[bazel] Add `final_symlink_outputs` to `bazel_action()` template.
This CL modifies the GN `bazel_action()` template to accept a new
argument, `final_symlink_outputs`, which is similar to `copy_outputs`
but is used to generate a symlink in the Ninja output directory that
points directly into the Bazel output_base (avoiding all copies).

This only makes sense for "final" Ninja targets that are never
used as input for other Ninja commands, as the timestamp of the
symlink target's is not maintained consistently by Bazel, except
if the output is a TreeArtifact.

This also adds a `ninja_base_dir` argument that can be used to
change the default base directory for `ninja` paths in output
scopes, which still default to `target_out_dir`.

+ Replaces //build/bazel:generate_fuchsia_in_tree_idk action()
  which was not properly incrementally correct, with the
  //build/bazel:final_fuchsia_in_tree_idk bazel_action().

+ Remove generate-fuchsia-idk-symlink.py which is no longer needed.

Bug: 367274537
Change-Id: I5db47eab9af5cf7279167662c2d70eee23531f42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143634
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>