[fidl][sdk] Remove support for "internal" SDK category in fidl()
For FIDL libraries, the "internal" SDK category has been equivalent to
not specifying a category in nearly all cases.

`fuchsia.boot`, which is used by C++ atoms in the "internal" SDK
category and must be in a category at least as visible as those atoms,
is now the only remaming use of "internal" in a `fidl()` library.

To prevent new uses of `sdk_category = "internal"` in instances of
`fidl()` and provide more helpful error messages, replace "internal"
with "internal_only_idk". Add an allowlist to prevent new uses.

Bug: 372986936, 333907192
Change-Id: Ic84b786ba1f2b17eb32c70f032398fe915e80e8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139638
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>