[netstack3] Allow conflict of identical connections
Currently, when the netstack attempts to finalize a connection that
conflicts with an existing entry in the conntrack table, finalize fails
and the packet is dropped. Usually this is not a problem because packets
that are part of the same flow are typically handled sequentially.
However, it's possible for two packets from the same flow to race, e.g.
if they are arriving on different interfaces. In this case they can both
end up creating an exclusive conntrack connection and then whichever
packet loses the race to finalize will be dropped.

To mitigate this scenario, when detecting conflicts, only report a
conflict (and drop the packet) if there is a conflicting connection that
is not identical to the one being finalized. In the case that the
connection to be finalized is identical to the one in the table (in
terms of both its original and reply tuples, *and* its NAT
configuration), adopt the already-finalized connection from the table
and allow the packet through.

Fixed: 375012282
Test: netstack3-filter-test
Multiply: netstack-socket-integration-test
Change-Id: I9530b1e6e987cabbc73067d4e2b2fc077d743a0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143973
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>