[usb][cdc] Fix pre-DdkAdd crash on bad parent
Right now, if the driver cannot connect to its upstream resources after
initializing its dispatcher but before handing the driver off to the DF
(i.e. any failure in Bind() post-dispatcher creation), the unique_ptr
will tear down the driver instance causing a crash due to kShutdown
check in the dispatcher.

This moves the setup logic to DdkInit(), which fixes the ordering of
operations such that if the something fails to initialize correctly,
DdkRelease() will get invoked (by the DF) cleaning up the dispatcher.

This prevents a crash in the failure-to-bind case.

Change-Id: I923413e0f4a588319308052ee936fac210f23876
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140355
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>