[ffx][usb_fastboot_discovery]: skip version check
Now opening an interface does not require the fastboot version variable to be 0.4 exactly.

As long as the getvar command returns either an Ok(Reply::Ok) or Ok(Reply::Fail) it is valid. It implies that the fastboot protocol is up and running and ready to communicate.

Bug: b/355707967
Fixed: b/355707967
Change-Id: I46c06fb876c67345209b26b999e1bd17d3d8b98c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141152
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>