[fdomain] Make futures disposable in client
This sets it so that most operations on handles that require an `await`
will still complete even if you throw the future away, so long as you're
polling the client's main future that was created when the client was
created.

We still require awaits for creating new sockets and channels and a few
other "global" operations.

Change-Id: I790fe26e35794634b4ab17d4bc529f6f65be6eba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121665
Reviewed-by: Steven Grady <slgrady@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>